Estimate CPS backscatter in Echoview CSV files

Estimate CPS backscatter in Echoview CSV files

extract_cps_nasc(
  path.in,
  pattern.in,
  path.out = path.in,
  suffix.out = "_nasc_cps.csv",
  x11.w = 1600,
  x11.h = 600,
  jpeg = TRUE,
  path.img = NULL,
  pattern.img = NULL,
  dist.bin = 2000,
  max.range = 350,
  transparency = 0.2,
  root = 1.5,
  scaling = 0.1,
  expansion = 2
)

Arguments

path.in

Directory containing CSV files.

pattern.in

Regular expression pattern used to list available CSV files.

path.out

Directory where outputs will be stored. Usually the same as path.in.

suffix.out

Suffix to append to resulting CSV file.

x11.w

Width of graphics window for line picks. Default = 1600 pixels.

x11.h

Height of graphics window for line picks. Default = 600 pixels.

jpeg

Save intermediate plots from line picks (TRUE/FALSE)?

path.img

Directory containing exported echogram images.

pattern.img

Regular expression pattern used to open corresponding echogram image.

dist.bin

Distance bin (m) for plotting results. Default is 2000 m, but should be made smaller for short transects.

max.range

Maximum depth range for bubble plots.

root

Constant for controlling bubble size. Default = 2 and should probably not be changed.

scaling

Constant for controlling bubble size. Default = 0.1 and should probably not be changed.

expansion

Constant for expanding axes in bubble plots.

Value

No objects are returned, but a new CSV file is produced along with several plots images showing the processing results.